[add] GlobalContext
